+++ app/src/main/java/pl/dkd/movieapp/domain/movies/model/MovieDetails.kt
@@ -1,3 +1,3 @@
 package pl.dkd.movieapp.domain.movies.model
 
-data class MovieDetails(val budget: Int, val revenue: Int)
+data class MovieDetails(val budget: Long, val revenue: Long)
